+++ CHANGELOG.md
@@ -26,6 +26,8 @@ Template for next version
 
 ### Fixed
 
+- field sort in excel submission export (https://github.com/ohmyform/ohmyform/issues/163)
+
 ### Security
 
 ## [1.0.1] - 2022-03-01